requirements.txt
Flask>=2.3
Flask-WTF>=1.1.1
WTForms>=3.1
pymongo>=4.6
Flask-Assets>=2.0
libsass>=0.22
openpyxl>=3.1
Werkzeug>=2.3
python-dateutil==2.9.0.post0
waitress>=2.1